Update store structure
