+++ src/store-easy-peasy/slices/contracts/helpers/viewFunction.ts
