+++ src/store-easy-peasy/slices/contracts/multisig/thunks/getMultisigContract.ts
