+++ src/store-easy-peasy/slices/contracts/multisig/thunks/getVersion.ts
