+++ src/store-easy-peasy/slices/contracts/multisig/thunks/view/getConfirmations.ts
@@ -1,5 +1,5 @@
 import { thunk } from "easy-peasy";
-import { viewFunction } from "~/store-easy-peasy/slices/multisig/helpers/viewFunction";
+import { viewFunction } from "~/store-easy-peasy/slices/contracts/helpers/viewFunction";
 
 export const getConfirmations = thunk(async (_, payload: any) => {
   const { contractId, provider, requestId }: any = payload;
