+++ src/store-easy-peasy/slices/contracts/multisig/thunks/view/getNumConfirmations.ts
@@ -1,5 +1,5 @@
 import { thunk } from "easy-peasy";
-import { viewFunction } from "~/store-easy-peasy/slices/multisig/helpers/viewFunction";
+import { viewFunction } from "~/store-easy-peasy/slices/contracts/helpers/viewFunction";
 
 export const getNumConfirmations = thunk(async (_, payload: any) => {
   const { contractId, provider }: any = payload;
