+++ src/store-easy-peasy/slices/contracts/multisig/thunks/view/listRequestIds.ts
@@ -1,5 +1,5 @@
 import { thunk } from "easy-peasy";
-import { viewFunction } from "~/store-easy-peasy/slices/multisig/helpers/viewFunction";
+import { viewFunction } from "~/store-easy-peasy/slices/contracts/helpers/viewFunction";
 
 export const listRequestIds = thunk(async (_, payload: any) => {
   const { contractId, provider }: any = payload;
