+++ src/store-easy-peasy/store.ts
@@ -2,7 +2,7 @@ import { createStore } from "easy-peasy";
 import { accounts } from "~/store-easy-peasy/slices/accounts";
 import { wallets } from "./slices/wallets";
 import { pages } from "src/store-easy-peasy/slices/pages";
-import { multisig } from "src/store-easy-peasy/slices/multisig";
+import { multisig } from "src/store-easy-peasy/slices/contracts/multisig";
 
 const model = {
   // nested slices
